Move `Qubit` and `Register` into separate files, add `Labeled` to `Qubit`
